requirements.txt
streamlit>=1.32.0
plotly>=5.18.0
pandas>=2.1.0
anthropic>=0.25.0
google-cloud-bigquery>=3.17.0
python-dotenv>=1.0.0
requests>=2.31.0
db-dtypes>=1.2.0
